pyproject.toml
[project]
name = "Better-Tools"
version = "0.0.6-beta"
description = "WebView with Streamlit Desktop Application"
readme = "README.md"
requires-python = "==3.13.*"
dependencies = [
    "pywebview>=6.1",
    "streamlit>=1.55.0",
    "fastapi>=0.135.2",
    "uvicorn>=0.42.0",
    "loguru>=0.7.3",
    "requests>=2.32.5",
    "xlwings>=0.34.0",
    "photoshop-python-api>=0.24.2",
    "pillow>=12.1.1",
    "tomli-w>=1.2.0",
    # "streamlit-cookies-controller>=0.0.4",
]

[project.optional-dependencies]
process = ["psutil>=7.2.2"]

[dependency-groups]
dev = [
    "pyinstaller>=6.17.0",
]
[[tool.uv.index]]
# url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple"
url = "https://mirrors.aliyun.com/pypi/simple"
default = true
# [tool.uv]
# python-install-mirror = "https://registry.npmmirror.com/-/binary/python-build-standalone/"